flask-api/requirements.txt
Flask==2.0.1
nextcloud-api-wrapper==0.2.3
werkzeug==2.0.3
python-dotenv
